backend/requirements.txt
Django==5.1.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
django-allauth==0.61.1
drf-yasg==1.21.7
django-request-logging==0.7.5
django-redis==5.4.0
redis==5.2.1
django-phonenumber-field[phonenumbers]

requests

asgiref


# # 1. Navigate to the project directory
# cd tutorclone_backend

# # 2. Install dependencies
# pip install -r requirements.txt

# # 3. Apply migrations
# python manage.py makemigrations
# python manage.py migrate

# # 4. Create superuser (optional)
# python manage.py createsuperuser

# # 5. Start Redis server (required for WebSocket)
# # On macOS: brew services start redis
# # On Linux: sudo service redis-server start
# # On Windows: Start Redis using Windows Service or Redis CLI

# # 6. Run the development server
# python manage.py runserver
